yisin: from develop to productdetail

+++ g-fashion-client/src/main/java/com/gfashion/restclient/magento/MagentoProductLinks.java
@@ -5,13 +5,16 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.json.JSONObject;
 
+import java.util.List;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class MagentoProductLinks {
+
     private String sku;
     private String name;
-    private Integer price;
+    private double price;
     private String image;
     private String link_type;
     private String linked_product_sku;
